+++ src/structured.jl
@@ -30,61 +30,18 @@ to the unit index.
 
 """ StructuredArray
 
-# Constructors for structured vectors and matrices.
-for (A, N) in ((:Vector, 1), (:Matrix, 2))
-    @eval begin
-        $(Symbol("Structured",A))(args...) = $(Symbol("Structured",A))(IndexCartesian, args...)
-        function $(Symbol("Structured",A))(::Union{S,Type{S}}, func, args...) where {S<:ConcreteIndexStyle}
-            T = guess_eltype(func, S, Val($N))
-            return StructuredArray{T,$N,S}(func, args...)
-        end
-        $(Symbol("Structured",A)){T}(args...) where {T} =
-            StructuredArray{T,$N,IndexCartesian}(args...)
-        $(Symbol("Structured",A)){T}(::Union{S,Type{S}}, args...) where {T,S<:ConcreteIndexStyle} =
-            StructuredArray{T,$N,S}(args...)
-        $(Symbol("Structured",A)){T,S}(args...) where {T,S<:ConcreteIndexStyle} =
-            StructuredArray{T,$N,S}(args...)
-    end
+# The following is the only outer constructor that calls the inner constructor. It takes care of
+# checking and converting the array shape arguments.
+function StructuredArray{T,N,S}(func, inds::NTuple{N,AxisLike}) where {T,N,S<:ConcreteIndexStyle}
+    check_shape(inds)
+    return StructuredArray{T,N,S}(BareBuild(), func, as_shape(inds))
 end
 
-# For structured arrays, the index style may be specified as the first argument.
-let cls = :StructuredArray
-    # 0-dimensional case.
-    @eval begin
-        $cls(       ::Union{S,Type{S}}, func) where {  S<:ConcreteIndexStyle} = $cls(    S, func, ())
-        $cls{T}(    ::Union{S,Type{S}}, func) where {T,S<:ConcreteIndexStyle} = $cls{T}( S, func, ())
-        $cls{T,0}(  ::Union{S,Type{S}}, func) where {T,S<:ConcreteIndexStyle} = $cls{T,0,S}(func, ())
-        $cls{T,0,S}(::Union{S,Type{S}}, func) where {T,S<:ConcreteIndexStyle} = $cls{T,0,S}(func, ())
-        $cls{T,0,S}(                    func) where {T,S<:ConcreteIndexStyle} = $cls{T,0,S}(func, ())
-    end
-    # N-dimensional cases with N ≥ 1.
-    for type in (:(Vararg{Union{Integer,AbstractRange{<:Integer}},N}),
-                 :(NTuple{N,Union{Integer,AbstractRange{<:Integer}}}))
-        @eval begin
-            $cls(       ::Union{S,Type{S}}, func, args::$type) where {  N,S<:ConcreteIndexStyle} = $cls(    S, func, to_inds(args))
-            $cls{T}(    ::Union{S,Type{S}}, func, args::$type) where {T,N,S<:ConcreteIndexStyle} = $cls{T}( S, func, to_inds(args))
-            $cls{T,N}(  ::Union{S,Type{S}}, func, args::$type) where {T,N,S<:ConcreteIndexStyle} = $cls{T,N,S}(func, to_inds(args))
-            $cls{T,N,S}(::Union{S,Type{S}}, func, args::$type) where {T,N,S<:ConcreteIndexStyle} = $cls{T,N,S}(func, to_inds(args))
-            $cls{T,N,S}(                    func, args::$type) where {T,N,S<:ConcreteIndexStyle} = $cls{T,N,S}(func, to_inds(args))
-        end
-    end
-    # Default index style is Cartesian for StructuredArray.
-    @eval begin
-        $cls(     func, args::Inds{N}) where {  N} = $cls(     IndexCartesian, func, args)
-        $cls{T}(  func, args::Inds{N}) where {T,N} = $cls{T,N}(IndexCartesian, func, args)
-        $cls{T,N}(func, args::Inds{N}) where {T,N} = $cls{T,N}(IndexCartesian, func, args)
-    end
-end
-
-# Constructors that manage to call the inner constructors.
-function StructuredArray(::Union{S,Type{S}}, func, args::Inds{N}) where {N,S<:ConcreteIndexStyle}
+# If element type `T` is unspecified, all constructors call the following one.
+function StructuredArray(::Union{S,Type{S}}, func, inds::NTuple{N,AxisLike}) where {N,S<:ConcreteIndexStyle}
     T = guess_eltype(func, S, Val(N))
-    return StructuredArray{T,N,S}(func, args)
+    return StructuredArray{T,N,S}(func, inds)
 end
-StructuredArray{T}(::Union{S,Type{S}}, func, args::Inds{N}) where {T,N,S<:ConcreteIndexStyle} =
-    StructuredArray{T,N,S}(func, args)
-StructuredArray{T,N}(::Union{S,Type{S}}, func, args::Inds{N}) where {T,N,S<:ConcreteIndexStyle} =
-    StructuredArray{T,N,S}(func, args)
 
 guess_eltype(func, ::Type{IndexLinear}, ::Val{N}) where {N} =
     Base.promote_op(func, Int)
@@ -92,10 +49,40 @@ guess_eltype(func, ::Type{IndexLinear}, ::Val{N}) where {N} =
 guess_eltype(func, ::Type{IndexCartesian}, ::Val{N}) where {N} =
     Base.promote_op(func, ntuple(Returns(Int), Val(N))...)
 
+# Other outer constructors for `StructuredArray` objects.
+for I in (:(NTuple{N,AxisLike}), :(Vararg{AxisLike,N}))
+    @eval begin
+        # Default index style is `IndexCartesian` for `StructuredArray` objects.
+        StructuredArray(     func, inds::$I) where {  N} = StructuredArray(    IndexCartesian, func, inds)
+        StructuredArray{T}(  func, inds::$I) where {T,N} = StructuredArray{T,N,IndexCartesian}(func, inds)
+        StructuredArray{T,N}(func, inds::$I) where {T,N} = StructuredArray{T,N,IndexCartesian}(func, inds)
+        # Constructors with specified element type `T` and index style `S`.
+        StructuredArray{T}(    ::Union{S,Type{S}}, func, inds::$I) where {T,N,S<:ConcreteIndexStyle} = StructuredArray{T,N,S}(func, inds)
+        StructuredArray{T,N}(  ::Union{S,Type{S}}, func, inds::$I) where {T,N,S<:ConcreteIndexStyle} = StructuredArray{T,N,S}(func, inds)
+        StructuredArray{T,N,S}(::Union{S,Type{S}}, func, inds::$I) where {T,N,S<:ConcreteIndexStyle} = StructuredArray{T,N,S}(func, inds)
+    end
+    I.args[1] === :Vararg || continue
+    @eval begin
+        # These constructors are defined elsewhere when shape is specified as a tuple.
+        StructuredArray(::Union{S,Type{S}}, func, inds::$I) where {  N,S<:ConcreteIndexStyle} = StructuredArray(    S, func, inds)
+        StructuredArray{T,N,S}(             func, inds::$I) where {T,N,S<:ConcreteIndexStyle} = StructuredArray{T,N,S}(func, inds)
+    end
+end
+
+# Constructors for structured vectors and matrices.
+for (A, N) in ((:Vector, 1), (:Matrix, 2)), I in (Vararg{AxisLike,N}, NTuple{N,AxisLike})
+    @eval begin
+        $(Symbol(:Structured,A))(                       func, inds::$I)                                 = StructuredArray(      func, inds)
+        $(Symbol(:Structured,A))(   ::Union{S,Type{S}}, func, inds::$I) where {  S<:ConcreteIndexStyle} = StructuredArray(   S, func, inds)
+        $(Symbol(:Structured,A)){T}(                    func, inds::$I) where {T}                       = StructuredArray{T}(   func, inds)
+        $(Symbol(:Structured,A)){T}(::Union{S,Type{S}}, func, inds::$I) where {T,S<:ConcreteIndexStyle} = StructuredArray{T}(S, func, inds)
+    end
+end
+
 # See comments near `getindex` in `abstractarray.jl` for explanations about how `getindex`
-# and `setindex!` methods are expected to be specialized depending on the indexing style:
-# `i:Int` for linear indexing style or `i::Vararg{Int,N}` for `N` dimensional Cartesian
-# indexing style.
+# `getindex` and `setindex!` methods are expected to be specialized depending on the
+# indexing style: `i:Int` for linear indexing style or `i::Vararg{Int,N}` for `N`
+# dimensional Cartesian indexing style.
 
 Base.IndexStyle(::Type{<:AbstractStructuredArray{T,N,S}}) where {T,N,S} = S()